Create README.md
A research proposal on the investigation of the modifications that current natural gas pipelines would require to transport hydrogen, and the mechanism of hydrogen penetration of current natural gas pipelines. The study is significant as it would facilitate the adoption of existing natural gas pipelines for hydrogen transportation; it would, thus, save the cost that would, otherwise, be incurred in constructing new hydrogen networks for hydrogen transmission and distribution as nations make plans and advocate for hydrogen as a major energy resource in the heat of the energy transition.